requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch==2.7.1+cu118

torch-geometric>=2.3.0
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.2.0
tqdm>=4.65.0
matplotlib>=3.7.0